# Request JS scripts with <link rel=preload...>