@CLAassistant 2016-03-12
[![CLA assistant check](https://cla-assistant.io/pull/badge/not_signed)](https://cla-assistant.io/redfin/react-server?pullRequest=60) <br/>You should sign our Contributor License Agreement in order to get your pull request merged.<br/>