@gigabo 2016-03-15
Very nice.

The only thing I worry about is the `PageConfig` toggle.  We'd like to be able to turn this on and off based on a value in an http response ("bouncer").  But `PageConfig` is synchronous, so we couldn't do that as this is currently set up.

I can think of a few other ways to toggle.  Unfortunately, I don't think any would be as clean as what you've got here:

- Add a lifecycle method to `PageUtil` (e.g. `getPreloadScripts`).  Could return a boolean, but a list of scripts would be more consistent with other page methods.  This would be pretty redundant with `getScripts` and `getSystemScripts`.
- Add a key to the `handleRoute` response object (e.g. `{code: 200, preloadScripts: true}`).  I'd rather not add a bunch of magical keys there.
- Add a key to the `script` objects returned by the `getScripts` and `getSystemScripts` lifecycle methods.  This would allow us to have a middleware that waits for our http response and modifies scripts on the way back up the stack.

That last one's not pretty, but maybe the least ugly of the bunch?

@roblg, @bartkusa, any better ideas?