@bartkusa 2016-03-15
Given that the goal is to tell the client ASAP what JS to start downloading, I don't think there's any value in waiting for async data to turn this on.

I was imagining that we'd enable this conditionally for some pages with a query param, eg, `?preload=1`. That'd let us measure it. If it looked good in controlled scenarios, we'd just hardcode it on for everybody on those pages.

I'm struggling to think of a scenario where this is only valuable sometimes. It seems like it's either always valuable or never valuable.