@gigabo 2016-03-15
> I don't think there's any value in waiting for async data to turn this on.

I place a pretty high value on the ability to dynamically control features at runtime using the aforementioned "bouncer" endpoint.

> the goal is to tell the client ASAP what JS to start downloading

To be clear, this is async data that the preloads will be waiting on _anyway_ (at Redfin) since `handleRoute` won't resolve until it's available.

> I'm struggling to think of a scenario where this is only valuable sometimes.

How about a script that I don't care about loading quickly?  Presumably omitting that from preload would free up resources for other more important scripts?  We currently have a single `LABjs` chain, but we've been talking for a while about supporting branches.  I don't want a preload from a secondary branch preempting a script in my critical path!