@doug-wade 2016-03-17
Talked with @bartkusa offline and ran some WPT tests against our details pages (specifically against [this test page](https://www.redfin.com/WA/Seattle/405-Prospect-St-98109/unit-102/home/2076248) ) and got the following results from [WebPageTest](http://www.webpagetest.org/)

Without preload

Performance Results (Median Run)
Document Complete	Fully Loaded
Load Time	First Byte	Start Render	User Time	Speed Index	DOM Elements	Time	Requests	Bytes In	Time	Requests	Bytes In
First View (Run 10)	10.628s	0.306s	1.894s	10.233s	2038	3997	10.628s	157	4,096 KB	11.185s	164	4,119 KB


With preload
Performance Results (Median Run)
Document Complete	Fully Loaded
Load Time	First Byte	Start Render	User Time	Speed Index	DOM Elements	Time	Requests	Bytes In	Time	Requests	Bytes In
First View (Run 8)	10.401s	0.330s	1.594s	9.825s	2099	4094	10.401s	157	4,102 KB	14.859s	165	4,125 KB
Plot Full Results 

Which seems to imply that we're gonna improve our page load by 220 millis.