@gigabo 2016-03-17
200ms would be huge! :rocket: 

Still would like an interface that supports asynchronous opt-in and per-script opt-in. ;)